auth_app/requirements.txt
bcrypt
twilio
streamlit
pyfingerprint
python-dotenv
PyJWT
pyserial
